authorization change implementation

+++ src/main/java/com/tms/Service/TMSApplicationService.java
@@ -6,6 +6,11 @@ import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.access.prepost.PostAuthorize;
+import org.springframework.security.access.prepost.PostFilter;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.access.prepost.PreFilter;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -17,6 +22,7 @@ import com.tms.exception.ErrorModel;
 import com.tms.exception.IErrorMessageConstants;
 import com.tms.model.Authority;
 import com.tms.model.TMSUser;
+import com.tms.model.TMSUser.Role;
 import com.tms.model.Task;
 import com.tms.repository.IAuthorityRepository;
 import com.tms.repository.ITaskRepository;
@@ -37,29 +43,47 @@ public class TMSApplicationService {
 	@Transactional
 	public Task createTask(Task taskTo) {
 		taskTo.setTaskId(null);
-		validateTaskOwner(taskTo);
+		TMSUser userDetails = getUserByEmail(SecurityContextHolder.getContext().getAuthentication().getName());
+		taskTo.setTaskOwner(userDetails);
 		return taskRepository.save(taskTo);
 	}
 
-	private void validateTaskOwner(Task taskTo) {
-		List<ErrorModel> errors = new ArrayList<>();
-		if (taskTo.getTaskOwner() != null && taskTo.getTaskOwner().getUserId() != null) {
-			TMSUser userDetails = getUser(taskTo.getTaskOwner().getUserId());
-			if (userDetails != null)
-				taskTo.setTaskOwner(userDetails);
-			else
-				errors.add(new ErrorModel(HttpStatus.NOT_FOUND.value(), IErrorMessageConstants.USER_NOT_FOUND));
+	@PostFilter("filterObject!=null?filterObject.taskOwner.email  == authentication.name:true")
+	public List<Task> getAllTasks() {
+		List<Task> allTasks = taskRepository.findAll();
+		return allTasks;
+	}
+
+	@PostAuthorize("returnObject!=null?returnObject.taskOwner.email == authentication.name:true")
+	public Task getTask(Long taskId) {
+		Task task = null;
+		Optional<Task> optionalTask = taskRepository.findById(taskId);
+		if (optionalTask.isPresent()) {
+			task = optionalTask.get();
 		}
-		if (!errors.isEmpty())
-			throw new BusinessException(errors);
+		return task;
+	}
+
+	@Transactional
+	@PostAuthorize("returnObject.taskOwner.email == authentication.name")
+	public Task updateTaskDetails(Long taskId, Task taskInputTo) {
+		Task taskTOUpdate = getTask(taskId);
+		if (taskTOUpdate != null)
+			updateAndValidateTaskFields(taskInputTo, taskTOUpdate);
+		else
+			throw new BusinessException(
+					List.of(new ErrorModel(HttpStatus.NOT_FOUND.value(), IErrorMessageConstants.TASK_NOT_FOUND)));
+		return taskRepository.save(taskTOUpdate);
 	}
 
 	@Transactional
 	public void deleteTask(Long taskId) {
-
 		Task task = getTask(taskId);
 		List<ErrorModel> errors = new ArrayList<>();
 		if (task != null) {
+			if (!task.getTaskOwner().getEmail()
+					.equals(SecurityContextHolder.getContext().getAuthentication().getPrincipal()))
+				throw new AccessDeniedException("Access Denied");
 			taskRepository.delete(task);
 		} else {
 			errors.add(new ErrorModel(HttpStatus.NOT_FOUND.value(), IErrorMessageConstants.TASK_NOT_FOUND));
@@ -70,40 +94,7 @@ public class TMSApplicationService {
 
 	}
 
-	public Task getTask(Long taskId) {
-		Task task = null;
-		Optional<Task> optionalTask = taskRepository.findById(taskId);
-		if (optionalTask.isPresent()) {
-			task = optionalTask.get();
-		}
-		return task;
-	}
-
-	public List<Task> getAllTasks() {
-		return taskRepository.findAll();
-	}
-
-	public List<Task> getAllTaskForUser(String email) {
-		List<Task> tasks=null;
-		TMSUser user = getUserByEmail(email);
-		if(user!=null) {
-			tasks=user.getTasks();
-		}
-		return tasks;
-	}
-
-	@Transactional
-	public Task updateTaskDetails(Long taskId, Task taskInputTo) {
-		Task taskTOUpdate = getTask(taskId);
-		if (taskTOUpdate != null)
-			updateTaskFields(taskInputTo, taskTOUpdate);
-		else
-			throw new BusinessException(
-					List.of(new ErrorModel(HttpStatus.NOT_FOUND.value(), IErrorMessageConstants.TASK_NOT_FOUND)));
-		return taskRepository.save(taskTOUpdate);
-	}
-
-	private void updateTaskFields(Task taskInputTo, Task taskTOUpdate) {
+	public void updateAndValidateTaskFields(Task taskInputTo, Task taskTOUpdate) {
 		List<ErrorModel> errors = new ArrayList<>();
 		boolean isChanged = false;
 		if (taskTOUpdate != null) {
@@ -139,11 +130,18 @@ public class TMSApplicationService {
 
 	/************ USER FUNCTIONALTIES **************/
 
+	@PostAuthorize("hasRole('ADMIN')")
 	public List<TMSUser> getAllUsers() {
-		return userRepository.findAll();
+		List<TMSUser> allUsers = userRepository.findAll();
+		if (allUsers != null && !allUsers.isEmpty())
+			return allUsers;
+		else
+			throw new BusinessException(
+					List.of(new ErrorModel(HttpStatus.NOT_FOUND.value(), IErrorMessageConstants.USER_NOT_FOUND)));
 
 	}
 
+	@PostAuthorize("hasRole('ADMIN') or returnObject.email==authentication.name")
 	public TMSUser getUser(Long userId) {
 		TMSUser tmsUser = null;
 		Optional<TMSUser> optionalUser = userRepository.findById(userId);
@@ -158,6 +156,7 @@ public class TMSApplicationService {
 		validateUserEmail(user);
 		validateUserAuthorities(user);
 		user.setPassword(passwordEncoder.encode(user.getPassword()));
+		user.setRole(Role.ROLE_USER);
 		userRepository.save(user);
 		return user;
 	}
@@ -167,7 +166,6 @@ public class TMSApplicationService {
 		user.setAuthorities(authorities);
 	}
 
-	@Transactional
 	public TMSUser updateUser(Long userId, TMSUser userInput) {
 		TMSUser userToBeUpdated = this.getUser(userId);
 		if (userToBeUpdated != null)
@@ -189,6 +187,11 @@ public class TMSApplicationService {
 	private void validateUserUpdates(TMSUser userInput, TMSUser userToBeUpdated) {
 		boolean isChanged = false;
 		List<ErrorModel> errors = new ArrayList<>();
+		if (!(SecurityContextHolder.getContext().getAuthentication().getPrincipal().equals(userToBeUpdated.getEmail())
+				|| SecurityContextHolder.getContext().getAuthentication().getAuthorities().stream()
+						.anyMatch(grantedAuthority -> grantedAuthority.getAuthority().equals("ROLE_ADMIN")))) {
+			throw new AccessDeniedException("Access Denied");
+		}
 		if (userToBeUpdated != null) {
 			if (!userToBeUpdated.getEmail().equals(userInput.getEmail())) {
 				validateUserEmail(userInput);
@@ -200,10 +203,20 @@ public class TMSApplicationService {
 				isChanged = true;
 			}
 
-			if (!passwordEncoder.matches(userInput.getPassword(),
-					(String) SecurityContextHolder.getContext().getAuthentication().getCredentials())) {
-				userToBeUpdated.setPassword(passwordEncoder.encode(userToBeUpdated.getPassword()));
+			if (!passwordEncoder.matches(userInput.getPassword(), userToBeUpdated.getPassword())) {
+				userToBeUpdated.setPassword(passwordEncoder.encode(userInput.getPassword()));
 				isChanged = true;
+			}
+			if (SecurityContextHolder.getContext().getAuthentication().getAuthorities().stream()
+					.anyMatch(grantedAuthority -> grantedAuthority.getAuthority().equals("ROLE_ADMIN"))) {
+				if (!userInput.getRole().equals(userToBeUpdated.getRole())) {
+					userToBeUpdated.setRole(userInput.getRole());
+					isChanged = true;
+				}
+				if (!userInput.getAuthorities().equals(userToBeUpdated.getAuthorities())) {
+					userToBeUpdated.setAuthorities(userInput.getAuthorities());
+					isChanged = true;
+				}
 			}
 
 			if (isChanged == false) {
@@ -214,13 +227,19 @@ public class TMSApplicationService {
 			throw new BusinessException(errors);
 	}
 
+	private void validateRoleAndAUthorityUpdates(TMSUser userInput, TMSUser userToBeUpdated) {
+
+	}
+
 	@Transactional
 	public void deleteUser(Long userId) {
 
 		TMSUser user = this.getUser(userId);
-		if (user != null)
+		if (user != null) {
+			if (!user.getEmail().equals(SecurityContextHolder.getContext().getAuthentication().getPrincipal()))
+				throw new AccessDeniedException("Access Denied");
 			userRepository.delete(user);
-		else
+		} else
 			throw new BusinessException(
 					List.of(new ErrorModel(HttpStatus.NOT_FOUND.value(), IErrorMessageConstants.USER_NOT_FOUND)));
 	}
